+++ CHANGES.md
@@ -14,6 +14,8 @@
 -   Added logic to select CBC events as preferred event over Burst. FAR acts
     as tie breaker when groups for preferred event and new event match.
 
+-   BAYESTAR now adds GraceDb URLs of events to FITS headers.
+
 ## 0.0.12 (2017-06-28)
 
 -   Prevent receiving duplicate copies of LVAlert messages by unregistering